# backend/docs/swagger.yaml
definitions:
...
  internal_handler_image.CreateKanikoRequest:
    properties:
      archs:
        items:
          type: string
        type: array
      description:
        type: string
      image:
        type: string
      name:
        type: string
      packages:
        type: string
      requirements:
        type: string
      tag:
        type: string
      tags:
        items:
          type: string
        type: array
      template:
        type: string
    type: object
...
  internal_handler_image.UpdateProjectQuotaRequest:
    properties:
      size:
        type: integer
    required:
    - size
    type: object
...
info:
  contact: {}
  description: This is the API server for Orbit, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Orbit API
  version: 1.0.0
paths:
...
  /v1/admin/images/kaniko:
    get:
      consumes:
      - application/json
      description: 管理员获取所有镜像制作信息
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 管理员获取相关镜像的功能
      tags:
      - ImagePack
    post:
      consumes:
      - application/json
      description: 获取参数，生成变量，调用接口
      parameters:
      - description: 创建ImagePack
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.CreateKanikoRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 创建ImagePack CRD和数据库kaniko entity
      tags:
      - ImagePack
  /v1/admin/images/remove:
    post:
      consumes:
      - application/json
      description: 管理员模式下根据任务状态智能处理：若任务状态为 Finished/Failed/Canceled 则执行删除操作；否则执行取消操作
      parameters:
      - description: 镜像构建任务ID列表
        in: body
        name: idList
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 管理员模式下删除或取消镜像制作任务（批量）
      tags:
      - ImagePack
...
  /v1/images/credential:
    post:
      consumes:
      - application/json
      description: 获取参数，生成变量，调用接口
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 创建用户的harbor项目，并返回用户的harbor项目的凭证
      tags:
      - ImagePack
...
  /v1/images/get:
    get:
      consumes:
      - application/json
      description: 获取imagepackname，搜索到imagepack的模板信息
      parameters:
      - description: 获取ImagePack的name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 获取imagepack的模板信息
      tags:
      - ImagePack
  /v1/images/getbyname:
    get:
      consumes:
      - application/json
      description: 获取imagepackname，搜索到imagepack
      parameters:
      - description: 获取ImagePack的name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 获取imagepack的详细信息
      tags:
      - ImagePack
...
  /v1/images/harbor:
    get:
      consumes:
      - application/json
      description: 通过后端获取harbor的部署地址
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 获取harbor的部署地址
      tags:
      - ImagePack
...
  /v1/images/kaniko:
    get:
      consumes:
      - application/json
      description: 返回该用户所有的镜像构建数据
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 用户获取镜像构建信息
      tags:
      - ImagePack
    post:
      consumes:
      - application/json
      description: 获取参数，生成变量，调用接口
      parameters:
      - description: 创建ImagePack CRD & Kaniko entity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.CreateKanikoRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 创建ImagePack CRD和数据库Kaniko entity
      tags:
      - ImagePack
  /v1/images/podname:
    get:
      consumes:
      - application/json
      description: 根据ID获取镜像构建Pod名称和命名空间
      parameters:
      - description: 镜像构建任务ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 获取镜像构建Pod名称
      tags:
      - ImagePack
  /v1/images/quota:
    get:
      consumes:
      - application/json
      description: 获取用户的project的详细信息
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 获取用户project的信息
      tags:
      - ImagePack
    post:
      consumes:
      - application/json
      description: 传入int64参数，查找用户的project，并更新镜像存储的配额
      parameters:
      - description: 更新镜像的ID和存储大小
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/internal_handler_image.UpdateProjectQuotaRequest'
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 更新project的配额
      tags:
      - ImagePack
  /v1/images/remove:
    post:
      consumes:
      - application/json
      description: 根据任务状态智能处理：若任务状态为 Finished/Failed/Canceled 则执行删除操作；否则执行取消操作
      parameters:
      - description: 镜像构建任务ID列表
        in: body
        name: idList
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 删除或取消镜像制作任务（批量）
      tags:
      - ImagePack
...
  /v1/images/valid:
    post:
      consumes:
      - application/json
      description: 通过获取的镜像链接列表，遍历其中的链接，检查是否有效
      produces:
      - application/json
      responses: {}
      security:
      - Bearer: []
      summary: 检查镜像链接是否有效
      tags:
      - ImagePack
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"